add mesh component to render core
