+++ crates/thyllore-model-core/src/mesh/mod.rs
@@ -5,7 +5,7 @@ mod values;
 mod vertex;
 
 pub use attribute::{VertexAttribute, VertexAttributeId, VertexFormat};
-pub use interleave::VertexLayout;
+pub use interleave::{compute_vertex_layout, create_interleaved_buffer, VertexLayout};
 pub use mesh_data::{MeshData, PrimitiveTopology};
 pub use values::VertexAttributeValues;
 pub use vertex::{Vertex, VertexData};
