+++ src/ecs/resource/gizmo/constraint_gizmo.rs
@@ -1,18 +1,1 @@
-use crate::ecs::component::{LineMesh, RenderInfo};
-
-#[derive(Clone, Debug)]
-pub struct ConstraintGizmoData {
-    pub visible: bool,
-    pub wire_mesh: LineMesh,
-    pub wire_render_info: RenderInfo,
-}
-
-impl Default for ConstraintGizmoData {
-    fn default() -> Self {
-        Self {
-            visible: false,
-            wire_mesh: LineMesh::default(),
-            wire_render_info: RenderInfo::default(),
-        }
-    }
-}
+pub use thyllore_render_core::ConstraintGizmoData;
