+++ src/ecs/resource/gizmo/light_gizmo.rs
@@ -1,13 +1,1 @@
-use crate::ecs::component::{GizmoDraggable, GizmoPosition, GizmoSelectable, LineMesh, RenderInfo};
-
-#[derive(Clone, Debug, Default)]
-pub struct LightGizmoData {
-    pub mesh: LineMesh,
-    pub render_info: RenderInfo,
-    pub position: GizmoPosition,
-    pub selectable: GizmoSelectable,
-    pub draggable: GizmoDraggable,
-    pub drag_active: bool,
-    pub ray_to_model: LineMesh,
-    pub vertical_lines: LineMesh,
-}
+pub use thyllore_render_core::LightGizmoData;
