+++ src/vulkanr/resource/mod.rs
@@ -1,12 +1,8 @@
-pub mod buffer_registry;
-pub mod pipeline_storage;
 pub mod raytracing_data;
-
-pub use buffer_registry::GpuBufferRegistry;
-pub use pipeline_storage::PipelineStorage;
 
 pub use thyllore_vulkan_core::resource::*;
 pub use thyllore_vulkan_core::resource::{
-    auto_exposure_buffers, bloom_chain, buffer, dof_buffer, dynamic_buffer, gbuffer,
-    graphics_resource, hdr_buffer, image, mesh_buffer, offscreen, onion_skin_pass,
+    auto_exposure_buffers, bloom_chain, buffer, buffer_registry, dof_buffer, dynamic_buffer,
+    gbuffer, graphics_resource, hdr_buffer, image, mesh_buffer, offscreen, onion_skin_pass,
+    pipeline_storage,
 };
